+++ app/src/main/java/com/petpal/mungmate/MainActivity.kt
@@ -1,10 +1,19 @@
 package com.petpal.mungmate
 
+import android.app.Application
+import android.content.ContentValues.TAG
+import android.content.pm.PackageInfo
+import android.content.pm.PackageManager
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.util.Base64
+import android.util.Log
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
+import com.kakao.util.maps.helper.Utility
+import java.security.MessageDigest
+import java.security.NoSuchAlgorithmException
 
 class MainActivity : AppCompatActivity() {
     private lateinit var navController: NavController
@@ -14,8 +23,12 @@ class MainActivity : AppCompatActivity() {
         installSplashScreen()
         setContentView(R.layout.activity_main)
 
+//        var keyHash = Utility.getKeyHash(this)
+//        Log.d(TAG, "keyhash : $keyHash")
+
         val navHostFragment =
             supportFragmentManager.findFragmentById(R.id.fragmentContainerView) as NavHostFragment
         navController = navHostFragment.navController
     }
+
 }
