+++ app/src/main/res/navigation/fragment_nav_graph.xml
@@ -46,7 +46,7 @@
     </fragment>
     <fragment
         android:id="@+id/item_walk"
-        android:name="com.petpal.mungmate.WalkFragment"
+        android:name="com.petpal.mungmate.ui.walk.WalkFragment"
         android:label="fragment_walk"
         tools:layout="@layout/fragment_walk" />
     <fragment
@@ -152,7 +152,7 @@
 
     <fragment
         android:id="@+id/item_walk"
-        android:name="com.petpal.mungmate.WalkFragment"
+        android:name="com.petpal.mungmate.ui.walk.WalkFragment"
         android:label="fragment_walk"
         tools:layout="@layout/fragment_walk" />
     <fragment
